# target/classes/static/swagger.yaml
openapi: 3.0.3
info:
  title: RESTfull API Service
  description: This API is designed for processing a sequence of numbers
  version: "1.0"
  contact:
    name: Pavel Kaznacheev
servers:
  - url: http://localhost:8080
    description: Generated server url
tags:
  - name: REST API
    description: API for processing a sequence
paths:
  /api/{operation}:
    post:
      tags:
        - REST API
      summary: Return value
      description: Return a value based on the received file name from JSON and operation from URL
      operationId: returnValue
      parameters:
        - name: operation
          in: path
          description: Operation from URL
          required: true
          schema:
            type: string
            enum:
              - GET_MAX_VALUE
              - GET_MIN_VALUE
              - GET_MEDIAN_VALUE
              - GET_ARITHMETIC_VALUE
              - GET_INCREASE_SEQUENCE
              - GET_DECREASE_SEQUENCE
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/incomingMapWoutOperation'
        required: true
      responses:
        '200':
          description: Return the result value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resultMap'
            application/xml:
              schema:
                $ref: '#/components/schemas/resultMap'
        '400':
          description: Return error message and timestamp
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorMap'
            application/xml:
              schema:
                $ref: '#/components/schemas/errorMap'
  /api/request:
    post:
      tags:
        - REST API
      summary: Return value
      description: Return a value based on the received file name and operation from JSON
      operationId: returnValue_1
      requestBody:
        description: Information about file and operation
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InformationDTO'
        required: true
      responses:
        '200':
          description: Return the result value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/resultMap'
            application/xml:
              schema:
                $ref: '#/components/schemas/resultMap'
        '400':
          description: Return error message and timestamp
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorMap'
            application/xml:
              schema:
                $ref: '#/components/schemas/errorMap'
components:
  schemas:
    resultMap:
      type: object
      properties:
        max_value:
          type: string
          example: '108'
    errorMap:
      type: object
      properties:
        message:
          type: string
          example: Operation not found
        timestamp:
          type: string
          example: '2022-10-23T13:38:40.006+00:00'
    incomingMapWoutOperation:
      type: object
      properties:
        fileName:
          type: string
          example: C:/folder/10m.txt
    InformationDTO:
      type: object
      properties:
        filePath:
          type: string
          description: File path
          example: C:/folder/10m.txt
        operation:
          $ref: '#/components/schemas/Operation'
      description: Information about file and operation
    Operation:
      type: object
      description: Name of the operation
      example: GET_MAX_VALUE
      oneOf:
        - type: string
          enum:
            - GET_MAX_VALUE
            - GET_MIN_VALUE
            - GET_MEDIAN_VALUE
            - GET_ARITHMETIC_VALUE
            - GET_INCREASE_SEQUENCE
            - GET_DECREASE_SEQUENCE